split bloating impl file into multiple separate ones
