--- align/__init__.py
@@ -1,3 +1,0 @@
-"""aligning the HPS detector whether in simulation or real"""
-
-from ._align import *
