+++ align/_cli.py
@@ -1,0 +1,4 @@
+"""Module to hold CLI object"""
+import typer
+
+app = typer.Typer()
