+++ align/_millepede.py
@@ -1,7 +1,14 @@
-"""Submodule helping construct the steering file for pede"""
-import sys
+"""run the pede minimizer over the bin files generated by tracking"""
+
 import os
-import math
+from typing import List
+
+import typer
+
+import _cmd
+from _cfg import cfg
+from _parameter import Parameter
+from _cli import app
 
 def getBeamspotConstraints(parMap):
     s = '\n!Beamspot constraints\n'
@@ -325,3 +332,128 @@ def getMeasurementD0(parMap):
     
     return s
 
+
+@app.command()
+def millepede(
+        input_file : List[str],
+        to_float : List[str] = typer.Option(..., help='parameters to float'),
+        out_dir : str = typer.Option(os.getcwd(), help='directory to save output to'),
+        prefix : str = typer.Option('',help='prefix to attach to files'),
+        year : int = typer.Option(2019,
+            help='year data was taken, determines set of parameters'),
+        survey_constraints : bool = typer.Option(False,
+            help='UNTESTED apply constraints from survey'),
+        beamspot_constraints : bool = typer.Option(False,
+            help='UNTESTED apply beam spot constraints'),
+        constraint_file : str = typer.Option(None,
+            help='UNTESTED optional external constraint file for pede'),
+        previous_fit : str = typer.Option(None,
+            help='UNTESTED optional fille containting previous pede fit solution'),
+        subito : bool = typer.Option(False,
+            help='add the "-s" parameter to pede')
+        ) :
+    """
+    Run the pede minimizer over the bin files generated by tracking
+    """
+
+    # get parameters depending on year
+    param_map_file = os.path.join(cfg.cfg().data_dir,'param_maps','hpsSvtParamMap.txt')
+    if year == 2019 :
+        param_map_file = os.path.join(cfg.cfg().data_dir,'param_maps','hpsSvtParamMap_2019.txt')
+
+    parameters = Parameter.parse_map_file(param_map_file)
+
+    # update parameters from previous fit
+    if previous_fit is not None :
+        Parameter.parse_pede_res(previous_fit, destination = previous_fit, skip_nonfloat = False)
+
+    # define which parameters are floating
+    floating = []
+    for f in to_float :
+        idn = None
+        if f.isnumeric() :
+            # string is a number, assume it is the idn
+            idn = int(f)
+        else:
+            # look for sensor name
+            for probe_id, p in parameters.items() :
+                if p.name == f :
+                    idn = prob_id
+                    break
+
+            if idn is None :
+                raise ValueError(f'Parameter {f} not found in parameter map.')
+
+        if idn not in parameters :
+            raise ValueError(f'Parameter {idn} not found in parameter map.')
+
+        parameters[idn].float()
+
+    # build steering file for pede
+    pede_steering_file = os.path.realpath(os.path.join(out_dir,prefix+'steer.txt'))
+    with open(pede_steering_file,'w') as psf :
+        # write out input mille binary files
+        psf.write('CFiles\n')
+        # scan each entry provided on command line,
+        #  recursively entering subdiretories and including
+        #  all '*.bin' files found
+        for ipf in input_file :
+            ipf = os.path.realpath(ipf)
+            if os.path.isfile(ipf) and ipf.endswith('.bin') :
+                psf.write(ipf+'\n')
+            elif os.path.isdir(ipf) :
+                for root, dirs, files in os.walk(ipf) :
+                    for name in files :
+                        if name.endswith('.bin') :
+                            psf.write(os.path.join(root,name)+'\n')
+
+        # external constraint file
+        if constraint_file is not None :
+            psf.write('\n')
+            psf.write('!Constraint file\n')
+            psf.write(constraint_file+'\n')
+
+        # parameters that can float
+        psf.write('\nParameter\n')
+        for i, p in parameters.items() : 
+            psf.write(p.pede_format() + '\n')
+
+        # survey constraints
+        if survey_constraints :
+            psf.write('\n!Survey constraints tu\n')
+            for p, name in param_map.items() :
+                if p.module_number() == 0 :
+                    continue
+                if p.direction() == 'u' and p.type() == 't' :
+                    psf.write('\nMeasurement 0.0 %.3f\n' % survey_meas_tu)
+                    psf.write('%s 1.0\n' & p)
+            psf.write("\n\n")
+        
+        # apply beamspotConstraint (This I think is not correct)
+        if beamspot_constraints:
+            #f.write(buildSteering.getBeamspotConstraints(paramMap))
+            psf.write(buildSteering.getBeamspotConstraintsFloatingOnly(pars))
+            psf.write("\n\n")
+        
+        psf.write("\n\n")
+        # determine MP minimization settings
+        with open(cfg.cfg().pede_minimization) as minfile :
+            for line in minfile :
+                psf.write(line)
+
+    # run pede
+    for ext in ['res','eve','log','his'] :
+        _cmd.run(['rm','-f', f'millepede.{ext}'], 
+                cwd=cfg.cfg().scratch)
+    pede_cmd = [cfg.cfg().pede, pede_steering_file]
+    if subito :
+        pede_cmd.append('-s')
+    _cmd.run(pede_cmd, cwd=cfg.cfg().scratch)
+
+    # move output to destination directory
+    for ext in ['res','eve','log','his'] :
+        f=os.path.join(cfg.cfg().scratch,f'millepede.{ext}')
+        if os.path.isfile(f) :
+            shutil.copy2(f,os.path.join(out_dir,prefix+f'millepede.{ext}'))
+
+    return
